Intelicarding an AI no longer causes two sets of AI laws to be reported at round-end.
You can now view the laws of an Intelicarded AI, and initiate a wipe of its files, killing it. Further functionality, such as giving it the ability to talk over a radio, is planned.

It is no longer possible to use an Intelicard on an AI core that has already had its AI removed.